basestation/requirements.txt
spidev
py-opc
requests
pyserial
